Refactor so that all test code for weather anom is in tests/weather_anomaly
