+++ tests/weather_anomaly/manage_views/conftest.py
