+++ tests/weather_anomaly/manage_views/test_manage_views.py
